refactor oscar actor pool code structure
