+++ mars/oscar/backends/communication/dummy.py
@@ -17,10 +17,10 @@ import concurrent.futures as futures
 from typing import Any, Callable, Coroutine, Dict, Type
 from urllib.parse import urlparse
 
-from .....utils import implements, classproperty
 from .base import Channel, ChannelType, Server, Client
 from .core import register_client, register_server
 from .errors import ChannelClosed
+from ....utils import implements, classproperty
 
 DEFAULT_DUMMY_ADDRESS = 'dummy://0'
