+++ mars/oscar/backends/communication/socket.py
@@ -23,11 +23,11 @@ from hashlib import md5
 from typing import Any, Dict, Callable, Coroutine, Type
 from urllib.parse import urlparse
 
-from .....utils import implements, to_binary, classproperty
-from .....serialization import AioSerializer, AioDeserializer, deserialize
 from .base import Channel, ChannelType, Server, Client
 from .core import register_client, register_server
 from .utils import read_buffers, write_buffers
+from ....serialization import AioSerializer, AioDeserializer, deserialize
+from ....utils import implements, to_binary, classproperty
 
 
 class SocketChannel(Channel):
