+++ mars/oscar/backends/mars/__init__.py
@@ -13,4 +13,3 @@
 # limitations under the License.
 
 from .backend import MarsActorBackend
-from .pool import create_actor_pool
