+++ mars/oscar/backends/ray/tests/test_communication.py
@@ -4,7 +4,7 @@ import pytest
 
 from mars.tests.core import require_ray
 from ..communication import ChannelID, Channel, RayServer, RayClient
-from ...mars.router import Router
+from ...router import Router
 from .....utils import lazy_import
 
 ray = lazy_import('ray')
